+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ## August
 
+* August 19 - Removed developer Twitter handle from Open Graph tags #607
 * August 19 - Convert reusable `/admin` components to View Components #605
 * August 18 - Add basic admin user search #604
 * August 17 - Send admins notifications on subscription changes #600